refactor(commenting): move the write path out of state.ts
state.ts is documented as transient UI state, and commitCommentMutation
was the one thing in it that wasn't — it's the undo/redo policy for
writes, and it was the file's only reason to import getCommentingOptions.

Moving it into comment-mutations.ts alone would have made comment-store
and comment-mutations import each other, since putCommentRecords applies
the policy. So the two raw writes move across with it: comment-store.ts is
now typed reads and imports nothing local, comment-mutations.ts owns the
whole write path (policy, raw writes, verbs), and state.ts is only state.

No public API change: the api-report is byte-identical, only the module a
symbol is declared in has changed.
